[FIX] models: speed up use of ir_property fields
The current logic treats ir_property fields as pure function fields,
and so will invalidate any cached ir_property field values whenever
any record in any model is modified.  This leads to cache thrashing
and a consequent O(n^2) performance penalty whenever updating a set of
records that happen to interact with any ir_property fields.

There seems to be no clean way to fix this, since the current hybrid
object model of old-style and new-style fields provides no generic way
for an ir_property field to indicate that it is not a pure function
field.

As an interim measure until the conversion to new-style fields is
complete across the whole codebase, fix this problem by explicitly
excluding ir_property fields from the list of pure function fields.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>