[FIX] report: Allow barcodes to be generated without authentication
When wkhtmltopdf is run as a result of an XML-RPC call, there is no
session cookie available and hence the use of

  command_args.extend(['--cookie', 'session_id', request.session.sid])

does not work to allow wkhtmltopdf to access internal links requiring
authentication.  One result is that barcodes do not appear in reports
generated via XML-RPC.

The current user is not relevant for generating barcode images, since
all of the required information is already present in the HTTP GET
parameter list.  Fix by marking the /report/barcode route as
auth="public" rather than auth="user".

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>