[IMP] ir_cron: Allow scheduled actions on individual records
There is no way to specify an individual target record via
ir.cron.args in the new API, since the record must be part of the
recordset on which the method is called.

Provide a "res_id" field (following the pattern in ir.model,
ir.values, ir.attachement and others) to allow a target record to be
specified as part of the ir.cron record.

As an added bonus, the "model" and "res_id" field pair may now be used
to provide a One2many field in the target model.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>