[FIX] stock: Use appropriate lot quantity in recompute_remaining_qty()
The done_qtys flag is used to select whether quantities should be the
"done" or "todo" quantities, but is ignored when retrieving quantities
from stock.pack.operation.lot records, with the result that the
recompute_remaining_qty() method fails to unambiguously identify the
appropriate quants and so creates link records with no reserved quant.

Fix by using packlot.qty or packlot.qty_todo as applicable.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>